+++ packages/fluorflow/lib/src/navigation/navigator_observer.dart
